Implemenation of SrmLs that utilized ListDirecroryHandler http://rb.dcache.org/r/1079/

+++ modules/srm/src/org/dcache/srm/unixfs/Storage.java
@@ -1043,14 +1043,13 @@ public class Storage
     @Override
     public List<FileMetaData>
         listDirectory(SRMUser user, String directory, final boolean verbose,
-                      int offset, int count)
+                      long offset, long count)
         throws SRMException
     {
         String[] list = listDirectory(user, directory, null);
         List<FileMetaData> result = new ArrayList<FileMetaData>();
-
-        for (int i = offset; i < list.length && i < offset + count; i++) {
-            result.add(getFileMetaData(user, directory + "/" + list[i], false));
+        for (long i = offset; i < list.length && i < offset + count; i++) {
+            result.add(getFileMetaData(user, directory + "/" + list[(int)i], false));
         }
 
         return result;
